+++ react-components/src/components/Table/index.tsx
@@ -1,20 +1,28 @@
-import React, { useState, useRef, useMemo } from 'react';
+import React, { useState, useRef, useMemo, useLayoutEffect } from 'react';
 import classNames from 'classnames';
 import {
   CellContext,
   createColumnHelper,
   flexRender,
   getCoreRowModel,
+  getSortedRowModel,
   Row,
+  SortingState,
   useReactTable,
 } from '@tanstack/react-table';
 import { useVirtualizer } from '@tanstack/react-virtual';
 
 import { TableProps } from './model';
 import TableRow from './components/RowTable';
+import { Icon } from '../Icons';
+import { sortIconNames } from './sortIconNames';
 import { FilterType } from './model/enum/filter-type.enum';
+import { useLocalStorage } from '../../utils/useLocalStorage';
 
 import s from './style.module.scss';
+
+const STORAGE_KEY = '_table_settings';
+const RESIZER_ATTRIBUTE_NAME = 'resizer';
 
 export function Table<T>({
   dataSource,
@@ -28,15 +36,41 @@ export function Table<T>({
   height = '100vh',
   onClick,
   acrossLine = false,
+  localStorageKey,
   className,
   ...props
 }: TableProps<T>): React.ReactElement {
+  const storageKey = `${localStorageKey}${STORAGE_KEY}`;
   const columnHelper = createColumnHelper<T>();
+  const [sorting, setSorting] = useState<SortingState>([]);
+  const [columnsSettings, setColumnsSettings] = useLocalStorage(
+    storageKey,
+    columns,
+  );
+
+  useLayoutEffect(() => {
+    const mergedColumns = columnsSettings.map((item) => ({
+      ...item,
+      cellComponent: columns.find(({ name }) => name === item.name)
+        ?.cellComponent,
+    }));
+    setColumnsSettings(mergedColumns);
+  }, []);
+
   const tableColumns = useMemo(() => {
-    return columns.map(
-      ({ name, caption, cellComponent, propertyName, filterType, width }) => {
+    return columnsSettings.map(
+      ({
+        name,
+        caption,
+        cellComponent,
+        propertyName,
+        filterType,
+        width,
+        sortable,
+      }) => {
         return columnHelper.accessor(name as any, {
           id: name,
+          enableSorting: sortable,
           cell: (info: CellContext<T, number>) => {
             const infoValue = info.getValue();
             let dictValue = '';
@@ -61,7 +95,7 @@ export function Table<T>({
               result = value;
             }
 
-            return <span>{result}</span>;
+            return <span className={s.tdContent}>{result}</span>;
           },
           header: () => caption,
           size: width,
@@ -77,11 +111,15 @@ export function Table<T>({
     columns: tableColumns,
     state: {
       rowSelection,
+      sorting,
     },
+    onSortingChange: setSorting,
     enableRowSelection: true,
     enableMultiRowSelection: false,
     onRowSelectionChange: setRowSelection,
+    columnResizeMode: 'onChange',
     getCoreRowModel: getCoreRowModel(),
+    getSortedRowModel: getSortedRowModel(),
   });
 
   const parentRef = useRef<HTMLDivElement>(null);
@@ -108,22 +146,75 @@ export function Table<T>({
         className={classNames(s.table, className)}
         style={{
           ...cellStyles,
+          width: table.getCenterTotalSize(),
           height: virtualizer.getTotalSize(),
         }}
       >
         <thead className={classNames(sticky && s.sticky)}>
           {table.getHeaderGroups().map((headerGroup) => (
             <tr key={headerGroup.id}>
-              {headerGroup.headers.map((header) => (
-                <th key={header.id} style={{ width: header.getSize() }}>
-                  {header.isPlaceholder
-                    ? null
-                    : flexRender(
-                        header.column.columnDef.header,
-                        header.getContext(),
-                      )}
-                </th>
-              ))}
+              {headerGroup.headers.map((header) => {
+                const isSorted = header.column.getIsSorted();
+
+                return (
+                  <th
+                    key={header.id}
+                    onClick={(event: React.MouseEvent<HTMLElement>) => {
+                      if (
+                        (event.target as HTMLElement).dataset.type ===
+                        RESIZER_ATTRIBUTE_NAME
+                      )
+                        return;
+
+                      const handler = header.column.getToggleSortingHandler();
+                      handler?.(event);
+                    }}
+                    className={classNames(
+                      header.column.getCanSort() && s.isSortable,
+                    )}
+                    style={{
+                      width: header.getSize(),
+                    }}
+                  >
+                    {header.isPlaceholder ? null : (
+                      <div className={s.thContent}>
+                        {flexRender(
+                          header.column.columnDef.header,
+                          header.getContext(),
+                        )}
+                        {isSorted ? (
+                          <Icon
+                            name={sortIconNames.get(isSorted) ?? ''}
+                            className={s.thContentIcon}
+                          />
+                        ) : null}
+                        <div
+                          data-type={RESIZER_ATTRIBUTE_NAME}
+                          className={classNames(
+                            s.resizer,
+                            header.column.getIsResizing() &&
+                              s.resizerIsResizing,
+                          )}
+                          onMouseDown={header.getResizeHandler()}
+                          onTouchStart={header.getResizeHandler()}
+                          onMouseUp={() => {
+                            const newColumnsSettings = columns.map(
+                              ({ cellComponent, ...rest }) => ({
+                                ...rest,
+                                width:
+                                  headerGroup.headers
+                                    .find(({ id }) => id === rest.name)
+                                    ?.getSize() ?? rest.width,
+                              }),
+                            );
+                            setColumnsSettings(newColumnsSettings);
+                          }}
+                        />
+                      </div>
+                    )}
+                  </th>
+                );
+              })}
             </tr>
           ))}
         </thead>
@@ -131,6 +222,8 @@ export function Table<T>({
         <tbody>
           {virtualizer.getVirtualItems().map((virtualRow) => {
             const row = rows[virtualRow.index] as Row<T>;
+
+            // console.log('row', row.getVisibleCells());
 
             return (
               <TableRow