+++ react-components/src/components/Table/sortIconNames.ts
@@ -1,0 +1,4 @@
+export const sortIconNames: Map<string, string> = new Map();
+
+sortIconNames.set('asc', 'arrowUp');
+sortIconNames.set('desc', 'arrowDown');