+++ react-components/src/components/Theme/index.tsx
@@ -1,31 +1,36 @@
 import React, { FC, useEffect, useMemo, useState } from 'react';
+
 import { ThemeProps } from './model';
+import { useLocalStorage } from '../../utils/useLocalStorage';
 
 export interface ThemeContextData {
   theme: string;
-  toggleTheme: (selectedTheme: string) => void;
+  changeTheme: (selectedTheme: string) => void;
 }
 
 const ThemeContext = React.createContext<ThemeContextData>({
-  theme: 'baselight',
-  toggleTheme: () => {
-    throw new Error('ToggleTheme not implemented');
+  theme: '',
+  changeTheme: () => {
+    throw new Error('changeTheme not implemented');
   },
 });
 
-const THEME_KEY = 'tetacom_theme_theme';
+const THEME_KEY = '_tetacom_theme';
 
-const ThemeContextProvider: FC<ThemeProps> = ({ defaultTheme, children }) => {
-  const savedValue: string =
-    localStorage.getItem(THEME_KEY) || defaultTheme || 'baselight';
+const ThemeContextProvider: FC<ThemeProps> = ({
+  defaultTheme = '',
+  localStorageKey,
+  children,
+}) => {
+  const storageKey = `${localStorageKey}${THEME_KEY}`;
+  const [savedValue, setValue] = useLocalStorage<string>(
+    storageKey,
+    defaultTheme || 'baselight',
+  );
 
   const [theme, setTheme] = useState<string>(savedValue);
   const DOMHtmlElement = useMemo(() => {
     return document.querySelector('html');
-  }, [theme]);
-
-  useEffect(() => {
-    localStorage.setItem(THEME_KEY, theme);
   }, [theme]);
 
   useEffect(() => {
@@ -40,8 +45,9 @@ const ThemeContextProvider: FC<ThemeProps> = ({ defaultTheme, children }) => {
   const themeContextValue: ThemeContextData = useMemo(() => {
     return {
       theme,
-      toggleTheme: (selectedTheme: string) => {
+      changeTheme: (selectedTheme: string) => {
         setTheme(selectedTheme);
+        setValue(theme);
       },
     };
   }, [theme]);