+++ react-components/src/components/Theme/model/index.ts
@@ -1,5 +1,9 @@
 import React from 'react';
 
 export interface ThemeProps extends React.PropsWithChildren {
+  // тема / название класса по умолчанию
   defaultTheme?: string;
+
+  // ключ в локальном хранилище
+  localStorageKey: string;
 }