+++ react-components/src/components/Theme/stories/Theme.stories.tsx
@@ -18,14 +18,16 @@ export default meta;
 
 type Story = StoryObj<typeof Provider>;
 
+const DEFAULT_THEME = 'baselight';
+
 const DefaultStory = () => (
-  <Provider>
+  <Provider localStorageKey="sb" defaultTheme={DEFAULT_THEME}>
     <Content />
   </Provider>
 );
 
 const Content = () => {
-  const { toggleTheme } = useContext(Context);
+  const { changeTheme } = useContext(Context);
 
   return (
     <Stack direction="column" align="start">
@@ -34,15 +36,15 @@ const Content = () => {
       </Typography.Title>
 
       <Select<BaseSelectProps>
-        value={{ key: 'tatneftlight', headline: 'Light' }}
+        value={{ key: DEFAULT_THEME, headline: 'Base Light' }}
         items={[
-          { key: 'baselight', headline: 'Base Light' },
-          { key: 'basedark', headline: 'Base Light' },
+          { key: DEFAULT_THEME, headline: 'Base Light' },
+          { key: 'basedark', headline: 'Base Dark' },
           { key: 'tatneftlight', headline: 'Tatneft Light' },
           { key: 'tatneftdark', headline: 'Tatneft Dark' },
         ]}
         onChangeItem={(item) => {
-          toggleTheme(item.key);
+          changeTheme(item.key);
         }}
       />